requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# client.py: 1,5
# nvidia.py: 1,2
discord.py == 0.16.12
